chore(triple-web-nextjs-pages): 폴더 구조 정리
