+++ packages/triple-web-nextjs-pages/src/initializers/index.ts
@@ -1,7 +1,10 @@
 import { NextPageContext } from 'next'
 
-import type { TripleWebProps } from './triple-web'
-import { getClientApp, getSession, getUserAgent } from './providers'
+import type { TripleWebProps } from '../providers/triple-web'
+
+import { getClientApp } from './client-app'
+import { getSession } from './session'
+import { getUserAgent } from './user-agent'
 
 export type BuildTripleWebPropsResult = Omit<
   TripleWebProps,
