--- packages/triple-web-nextjs-pages/src/providers/client-app.ts
@@ -1,38 +1,0 @@
-import type { NextPageContext } from 'next'
-import { ClientAppName, type ClientAppValue } from '@titicaca/triple-web'
-import { clientAppRegex, macAppRegex } from '@titicaca/triple-web-utils'
-
-export function getClientApp(ctx: NextPageContext): ClientAppValue {
-  const userAgent = ctx.req
-    ? (ctx.req.headers['user-agent'] ?? '')
-    : window.navigator.userAgent
-  const metadata = clientAppRegex.exec(userAgent)
-
-  if (!metadata) {
-    return null
-  }
-
-  const autoplay =
-    (ctx.req?.headers['x-triple-autoplay'] as
-      | NonNullable<ClientAppValue>['device']['autoplay']
-      | undefined) ?? 'always'
-  const networkType =
-    (ctx.req?.headers['x-triple-network-type'] as
-      | NonNullable<ClientAppValue>['device']['networkType']
-      | undefined) ?? 'unknown'
-
-  return {
-    metadata: {
-      name:
-        metadata[1] === 'Triple-Android'
-          ? ClientAppName.Android
-          : ClientAppName.iOS,
-      version: metadata[2],
-      isMacApp: macAppRegex.test(userAgent),
-    },
-    device: {
-      autoplay,
-      networkType,
-    },
-  }
-}
