+++ packages/triple-web-nextjs-pages/src/providers/event-metadata-provider.mdx
